editor/requirements.txt
contourpy==1.2.1
cycler==0.12.1
exceptiongroup==1.2.1
fonttools==4.52.1
iniconfig==2.0.0
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
pyparsing==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
scipy==1.13.0
six==1.16.0
tomli==2.0.1
